refactor(ui): ディレクトリ構成整理
機能単位ではなくUIで達成する事柄ベースに再配置
