+++ src/view/index/filter/range/CoamRangeSlider.svelte
@@ -3,12 +3,11 @@
   import {roundUpByRealPart} from "~core/utils/number.ts";
 
   import i18n from "~view/i18n/define.ts";
+  import RangeSlider from '~view/index/filter/range/base/RangeSlider.svelte'
 
   import type {Candidates} from "~data/types/candidates.ts";
 
   import {createEventDispatcher} from "svelte";
-
-  import RangeSlider from './base/RangeSlider.svelte'
 
   export let candidates: Candidates
