+++ src/view/index/filter/range/base/RangeSlider.svelte
